requirements.txt
flake8~=3.8.3
flake8-logging-format~=0.6.0
black~=19.10b0
isort[pyproject]~=4.3.21
mypy~=0.780
sqlalchemy~=1.3.18
mysqlclient~=2.0.1
request~=2.24.0
beautifulsoup4~=4.9.1